+++ components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/preferences/BrowserPrefObserverIntegration.kt
@@ -50,6 +50,28 @@ class BrowserPrefObserverIntegration(
         )
     }
 
+    /**
+     * Will register a list of browser preferences for observation. Will receive reports on changes via
+     * [onPreferenceChange].
+     *
+     * @param prefs The list of browser preferences to observe.
+     * @param onSuccess What to do in the event of a successful registration.
+     * @param onError What to do in the event of a unsuccessful registration.
+     */
+    fun registerPrefsForObservation(
+        prefs: List<String>,
+        onSuccess: () -> Unit,
+        onError: (Throwable) -> Unit,
+    ) {
+        engine.registerPrefsForObservation(
+            prefs,
+            onSuccess = {
+                onSuccess
+            },
+            onError = { onError },
+        )
+    }
+
 /**
  * Will unregister a browser preference for observation that was registered via [registerPrefForObservation].
  *
@@ -65,6 +87,29 @@ class BrowserPrefObserverIntegration(
     ) {
         engine.unregisterPrefForObservation(
             pref,
+            onSuccess = {
+                onSuccess
+            },
+            onError = { onError },
+        )
+    }
+
+    /**
+     * Will unregister a list of browser preferences for
+     * observation that was registered via [registerPrefForObservation].
+     *
+     * @param prefs The list of browser preferences to stop observing.
+     * @param onSuccess What to do in the event of a successful deregistration.
+     * @param onError What to do in the event of a unsuccessful deregistration.
+     *
+     */
+    fun unregisterPrefsForObservation(
+        prefs: List<String>,
+        onSuccess: () -> Unit,
+        onError: (Throwable) -> Unit,
+    ) {
+        engine.unregisterPrefsForObservation(
+            prefs,
             onSuccess = {
                 onSuccess
             },