+++ components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/preferences/BrowserPrefObserverIntegrationTest.kt
@@ -5,6 +5,7 @@ import mozilla.components.concept.engine.preferences.BrowserPreference
 import mozilla.components.support.test.any
 import mozilla.components.support.test.mock
 import org.junit.Test
+import org.mockito.ArgumentMatchers.anyList
 import org.mockito.ArgumentMatchers.anyString
 import org.mockito.Mockito.verify
 
@@ -48,6 +49,23 @@ class BrowserPrefObserverIntegrationTest {
     }
 
     @Test
+    fun `WHEN registerPrefsForObservation THEN the engine method is invoked`() {
+        val engine = mock<Engine>()
+        val feature = BrowserPrefObserverIntegration(engine)
+
+        val onSuccess: () -> Unit = {}
+        val onError: (Throwable) -> Unit = {}
+        feature.registerPrefsForObservation(
+            listOf("test.item"),
+            onSuccess,
+            onError,
+        )
+
+        verify(engine)
+            .registerPrefsForObservation(anyList<String>(), any(), any())
+    }
+
+    @Test
     fun `WHEN unregisterPrefForObservation THEN the engine method is invoked`() {
         val engine = mock<Engine>()
         val feature = BrowserPrefObserverIntegration(engine)
@@ -62,6 +80,23 @@ class BrowserPrefObserverIntegrationTest {
 
         verify(engine)
             .unregisterPrefForObservation(anyString(), any(), any())
+    }
+
+    @Test
+    fun `WHEN unregisterPrefsForObservation THEN the engine method is invoked`() {
+        val engine = mock<Engine>()
+        val feature = BrowserPrefObserverIntegration(engine)
+
+        val onSuccess: () -> Unit = {}
+        val onError: (Throwable) -> Unit = {}
+        feature.unregisterPrefsForObservation(
+            listOf("test.item"),
+            onSuccess,
+            onError,
+        )
+
+        verify(engine)
+            .unregisterPrefsForObservation(anyList<String>(), any(), any())
     }
 
     @Test