+++ docs/changelog.md
@@ -7,6 +7,8 @@ permalink: /changelog/
 # 149.0 (In Development)
 * **browser-engine-gecko**
   * 🆕 New Engine Settings API `firefoxRelay` that can be see to different modes with `FirefoxRelayMode`.
+* **browser-engine-gecko** and **concept-engine**
+  * Updated GeckoEngine and BrowserPreferencesRuntime to accommodate registering and unregistering multiple browser preferences at a time for observation. [Bug 2006095](https://bugzilla.mozilla.org/show_bug.cgi?id=2006095)
 
 # 148.0
 * **support-utils**